ODY-170 got rid of add block and updated message for if no droplet
