+++ frontend/components/draft/lesson/draggable_block_list_client.tsx
@@ -1,9 +1,12 @@
 "use client";
-
 import { useCallback } from "react";
-import { Block } from "./add-block";
+import { Block } from "./add-tools";
 import DraggableBlockTile from "./draggable_block_tile";
-
+import {
+  Message,
+  MessageDescription,
+  MessageHeader,
+} from "@/components/message";
 export function DraggableBlockListClient({
   blocks,
   moveCard,
@@ -25,25 +28,32 @@ export function DraggableBlockListClient({
     },
     [onAddBlock],
   );
-
+  if (blocks.length === 0) {
+    return (
+      <Message className="mb-8 rounded-md dark:border-slate-500 dark:bg-slate-800">
+        <MessageHeader subtitle="" title="No Blocks" />
+        <MessageDescription>
+          Use the add button to get started!
+        </MessageDescription>
+      </Message>
+    );
+  }
   return (
-    <>
-      <div className="grid grid-cols-1 space-y-8">
-        {blocks.map((block, index) => (
-          <div
-            key={`${block.__component}-${index}-${blocks.length}`}
-            className="flex w-full flex-col items-center justify-center"
-          >
-            <DraggableBlockTile
-              block={block}
-              index={index}
-              moveCard={moveCard}
-              setBlock={setBlock}
-              deleteBlock={deleteBlock}
-            />
-          </div>
-        ))}
-      </div>
-    </>
+    <div className="grid grid-cols-1 space-y-8">
+      {blocks.map((block, index) => (
+        <div
+          key={`${block.__component}-${index}-${blocks.length}`}
+          className="flex w-full flex-col items-center justify-center"
+        >
+          <DraggableBlockTile
+            block={block}
+            index={index}
+            moveCard={moveCard}
+            setBlock={setBlock}
+            deleteBlock={deleteBlock}
+          />
+        </div>
+      ))}
+    </div>
   );
 }
